requirements.txt
# =============================================================================
# SISTEMA DE LABORATORIOS SENA - DEPENDENCIAS UNIFICADAS
# =============================================================================
# Python 3.11+ Compatible
# Estado: ✅ PRODUCCIÓN READY
# Actualizado: 25 de Abril de 2026
# =============================================================================

# =====================
# FRAMEWORK WEB Y API
# =====================
Flask==3.1.1
Flask-RESTful==0.3.10
Flask-JWT-Extended==4.7.1
flask-cors==6.0.1
Flask-Limiter==3.5.0
Werkzeug==3.1.3

# =====================
# BASE DE DATOS
# =====================
mysql-connector-python==9.4.0

# =====================
# CONFIGURACIÓN Y ENTORNO
# =====================
python-dotenv==1.0.0
python-decouple==3.8

# =====================
# PROCESAMIENTO DE IMÁGENES Y VISIÓN COMPUTACIONAL
# =====================
opencv-python==4.12.0.88
Pillow==11.3.0
numpy==2.2.6

# =====================
# RECONOCIMIENTO DE VOZ Y SÍNTESIS
# =====================
SpeechRecognition==3.10.0
pyttsx3==2.90

# =====================
# UTILIDADES DE FECHA Y TIEMPO
# =====================
python-dateutil==2.9.0.post0
pytz==2025.2

# =====================
# MONITOREO Y SISTEMA
# =====================
psutil==7.1.0

# =====================
# UTILIDADES GENERALES
# =====================
requests==2.32.5
urllib3==2.5.0
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10

# =====================
# GENERACIÓN DE REPORTES
# =====================
reportlab==4.2.5
openpyxl==3.1.5
xlsxwriter==3.2.0

# =====================
# DESARROLLO Y UTILIDADES
# =====================
Faker==34.0.0

# =============================================================================
# DEPENDENCIAS OPCIONALES (COMENTADAS)
# =============================================================================
# Descomentar solo si se necesitan estas funcionalidades

# =====================
# RECONOCIMIENTO FACIAL (OPCIONAL)
# =====================
# Nota: face-recognition requiere dlib, que puede ser difícil de instalar en Windows
# Si tienes problemas, descarga wheels precompilados desde:
# https://github.com/z-mahmud22/Dlib_Windows_Python3.x
# face-recognition==1.3.0  # Requiere dlib - Instalar manualmente si se necesita
# dlib==19.24.6  # Descomentar después de instalar desde wheel

# =====================
# CAPTURA DE AUDIO (OPCIONAL)
# =====================
# Nota: En Windows, para captura de audio en tiempo real necesitas PyAudio
# Instalación recomendada:
#   pip install pipwin
#   pipwin install pyaudio
# O descarga wheel desde: https://www.lfd.uci.edu/~gohlke/pythonlibs/#pyaudio
# pyaudio==0.2.14

# =====================
# IA AVANZADA - RECONOCIMIENTO DE VOZ (OPCIONAL)
# =====================
# Nota: DeepSpeech está descontinuado, considera usar alternativas como:
# - Vosk (offline, ligero)
# - Whisper de OpenAI (más preciso)
# deepspeech==0.9.3
# webrtcvad==2.0.10  # Requiere compilación - OPCIONAL

# =====================
# IA AVANZADA - VISIÓN POR COMPUTADORA (OPCIONAL)
# =====================
# Nota: TensorFlow es pesado. Para producción considera tflite-runtime
# tensorflow==2.16.1
# tflite-runtime==2.14.0  # Alternativa ligera (solo inferencia)
# scikit-image==0.24.0  # Requiere compilación en Windows - OPCIONAL

# =====================
# UTILIDADES ADICIONALES PARA IA (OPCIONAL)
# =====================
# librosa==0.10.2  # Para procesamiento de audio avanzado
# matplotlib==3.9.0  # Para visualización
# seaborn==0.13.2  # Para gráficos estadísticos

# =============================================================================
# DEPENDENCIAS DE DESARROLLO (COMENTADAS)
# =============================================================================
# Descomentar solo para desarrollo y testing
# pytest==8.3.5
# black==25.1.0
# flake8==7.2.0

# =============================================================================
# NOTAS DE INSTALACIÓN
# =============================================================================
# 1. Instalación básica (producción):
#    pip install -r requirements_unificado.txt
#
# 2. Para reconocimiento facial (opcional):
#    - Descarga dlib wheel para Python 3.11 desde:
#      https://github.com/z-mahmud22/Dlib_Windows_Python3.x
#    - pip install dlib-19.24.6-cp311-cp311-win_amd64.whl
#    - pip install face-recognition
#
# 3. Para captura de audio (opcional):
#    - pip install pipwin
#    - pipwin install pyaudio
#
# 4. Para IA avanzada (opcional):
#    - pip install tensorflow  # O tflite-runtime para versión ligera
#    - pip install librosa matplotlib seaborn
#
# 5. El sistema funciona sin las dependencias opcionales,
#    usando implementaciones de respaldo con OpenCV
#
# 6. Para verificar instalación:
#    python check_setup.py
#
# 7. Para desarrollo con herramientas adicionales:
#    pip install -r requirements_unificado.txt --editable
#    pip install pytest black flake8
# =============================================================================
